FindTest and ToggleSeenValueTest for TicketService are added, CommentService tests are added

+++ src/main/java/io/inprice/api/app/ticket/CommentService.java
@@ -23,7 +23,7 @@ import io.inprice.common.models.TicketComment;
 public class CommentService {
 
 	Response insert(TicketCommentDTO dto) {
-		Response res = Responses.Invalid.TICKET;
+		Response res = Responses.Invalid.COMMENT;
 
 		String problem = validate(dto);
 		if (problem == null) {
@@ -48,8 +48,10 @@ public class CommentService {
 							res = Responses.DataProblem.DB_PROBLEM;
 						}
 					} else {
-						res = new Response("You are not allowed to add comment to a closed ticket!");
+						res = Responses.NotAllowed.CLOSED_TICKET;
 					}
+				} else {
+					res = Responses.NotFound.TICKET;
 				}
 			}
 		} else {
@@ -60,7 +62,7 @@ public class CommentService {
 	}
 
 	Response update(TicketCommentDTO dto) {
-		Response res = Responses.Invalid.TICKET;
+		Response res = Responses.NotFound.COMMENT;
 
 		if (dto.getId() != null && dto.getId() > 0) {
 			String problem = validate(dto);
@@ -71,7 +73,7 @@ public class CommentService {
 					Ticket ticket = ticketDao.findById(dto.getTicketId(), dto.getAccountId());
 					if (ticket != null) {
 						if (! TicketStatus.CLOSED.equals(ticket.getStatus())) {
-							TicketComment comment = ticketDao.findCommentById(dto.getId());
+							TicketComment comment = ticketDao.findCommentById(dto.getId(), dto.getAccountId());
 							if (comment != null) {
 								if (comment.getEditable()) {
 			  					if (CurrentUser.getRole().equals(UserRole.ADMIN) || comment.getUserId().equals(CurrentUser.getUserId())) {
@@ -92,8 +94,10 @@ public class CommentService {
 								}
 							}
 						} else {
-							res = new Response("You are not allowed to update comment on a closed ticket!");
+							res = Responses.NotAllowed.CLOSED_TICKET;
 						}
+					} else {
+						res = Responses.NotFound.TICKET;
 					}
 				}
 			} else {
@@ -104,39 +108,43 @@ public class CommentService {
 	}
 	
 	Response delete(Long id) {
-		Response res = Responses.NotFound.TICKET;
+		Response res = Responses.NotFound.COMMENT;
 
 		if (id != null && id > 0) {
 			try (Handle handle = Database.getHandle()) {
 				TicketDao ticketDao = handle.attach(TicketDao.class);
 
-				TicketComment comment = ticketDao.findCommentById(id);
+				TicketComment comment = ticketDao.findCommentById(id, CurrentUser.getAccountId());
 
 				if (comment != null) {
 					if (comment.getEditable()) {
   					Ticket ticket = ticketDao.findById(comment.getTicketId(), comment.getAccountId());
   
-  					if (! TicketStatus.CLOSED.equals(ticket.getStatus())) {
-  						if (CurrentUser.getRole().equals(UserRole.ADMIN) || comment.getUserId().equals(CurrentUser.getUserId())) {
-  							
-  							handle.begin();
-  							ticketDao.makeOnePreviousCommentEditable(comment.getTicketId(), id);
-  
-  							boolean isOK = ticketDao.deleteCommentById(id);
-  							if (isOK) {
-  								ticketDao.decreaseCommentCount(comment.getTicketId());
-  								handle.commit();
-									List<TicketComment> commentList = ticketDao.fetchCommentListByTicketId(comment.getTicketId());
-									res = new Response(commentList);
-  							} else {
-  								handle.rollback();
-  								res = Responses.DataProblem.DB_PROBLEM;
-  							}
-  						} else {
-  							res = Responses.PermissionProblem.WRONG_USER;
-  						}
+  					if (ticket != null) {
+    					if (! TicketStatus.CLOSED.equals(ticket.getStatus())) {
+    						if (CurrentUser.getRole().equals(UserRole.ADMIN) || comment.getUserId().equals(CurrentUser.getUserId())) {
+    							
+    							handle.begin();
+    							ticketDao.makeOnePreviousCommentEditable(comment.getTicketId(), id);
+    
+    							boolean isOK = ticketDao.deleteCommentById(id);
+    							if (isOK) {
+    								ticketDao.decreaseCommentCount(comment.getTicketId());
+    								handle.commit();
+  									List<TicketComment> commentList = ticketDao.fetchCommentListByTicketId(comment.getTicketId());
+  									res = new Response(commentList);
+    							} else {
+    								handle.rollback();
+    								res = Responses.DataProblem.DB_PROBLEM;
+    							}
+    						} else {
+    							res = Responses.PermissionProblem.WRONG_USER;
+    						}
+    					} else {
+  							res = Responses.NotAllowed.CLOSED_TICKET;
+    					}
 						} else {
-							res = new Response("You are not allowed to delete comment from a closed ticket!");
+  						res = Responses.NotFound.TICKET;
 						}
 					} else {
 						res = Responses.NotAllowed.UPDATE;
@@ -154,11 +162,11 @@ public class CommentService {
 		if (StringUtils.isBlank(dto.getBody())) {
 			problem = "Body cannot be empty!";
 		} else if (dto.getBody().length() < 12 || dto.getBody().length() > 1024) {
-			problem = "Body must be between 12-1024 chars!";
+			problem = "Body must be between 12 - 1024 chars!";
 		}
 
 		if (problem == null && dto.getTicketId() == null) {
-			problem = "Ticket not found!";
+			problem = "Ticket id cannot be empty!";
 		}
 
 		if (problem == null) {
